+++ apps/client/src/pages/LandingPage.tsx
@@ -1,16 +1,19 @@
+import { useQuery } from '@tanstack/react-query';
 import { useState } from 'react';
 
 import { SignInForBookmarkModal } from '@/components/modal/SignInForBookmarkModal';
 import { GlobalNavigationBar } from '@/components/nav/GlobarNavigationBar';
 import type { FilterElements, JobMinorCategory } from '@/entities/post';
+import type { Series } from '@/entities/post';
 import {
   FilterSection,
   NarrowRolesFilter,
   PaginationBar,
   PostCard,
   RolesFilter,
-  useGetPosts,
 } from '@/feature/landing';
+import { useGuardContext } from '@/shared/context/hooks';
+import { ServiceContext } from '@/shared/context/ServiceContext';
 import { useRouteNavigation } from '@/shared/route/useRouteNavigation';
 
 export const LandingPage = () => {
@@ -35,16 +38,12 @@ export const LandingPage = () => {
   const [currentPage, setCurrentPage] = useState(0);
   const [currentGroup, setCurrentGroup] = useState(0);
 
-  const { data: postsData } = useGetPosts({
+  const { postsData } = useGetPosts({
     page: currentPage,
     ...filterElements,
   });
 
-  if (postsData === undefined) {
-    return <p>로딩 중...</p>;
-  }
-
-  const TOTAL_PAGES = postsData.paginator.lastPage;
+  const TOTAL_PAGES = postsData?.paginator.lastPage;
   const PAGES_PER_GROUP = 5;
 
   return (
@@ -84,33 +83,36 @@ export const LandingPage = () => {
             {/* 게시글 리스트 */}
             <main>
               <div className="grid w-full grid-cols-1 md:grid-cols-2 xl:grid-cols-3 gap-6 pr-2">
-                {postsData.posts.map((post, idx) => (
-                  <PostCard
-                    key={`post-${idx}`}
-                    post={post}
-                    onDetailClick={(postId) => {
-                      toPost({ postId });
-                    }}
-                    setShowSignInModal={setShowSignInModal}
-                  />
-                ))}
+                {postsData !== undefined &&
+                  postsData.posts.map((post, idx) => (
+                    <PostCard
+                      key={`post-${idx}`}
+                      post={post}
+                      onDetailClick={(postId) => {
+                        toPost({ postId });
+                      }}
+                      setShowSignInModal={setShowSignInModal}
+                    />
+                  ))}
               </div>
             </main>
 
             {/* 페이지네이션 */}
             <footer className="mt-6 flex justify-center">
-              <PaginationBar
-                totalPages={TOTAL_PAGES}
-                pagesPerGroup={PAGES_PER_GROUP}
-                currentPage={currentPage}
-                currentGroup={currentGroup}
-                onChangePage={(page) => {
-                  setCurrentPage(page);
-                }}
-                onChangeGroup={(group) => {
-                  setCurrentGroup(group);
-                }}
-              />
+              {TOTAL_PAGES !== undefined && (
+                <PaginationBar
+                  totalPages={TOTAL_PAGES}
+                  pagesPerGroup={PAGES_PER_GROUP}
+                  currentPage={currentPage}
+                  currentGroup={currentGroup}
+                  onChangePage={(page) => {
+                    setCurrentPage(page);
+                  }}
+                  onChangeGroup={(group) => {
+                    setCurrentGroup(group);
+                  }}
+                />
+              )}
             </footer>
           </div>
         </div>
@@ -118,4 +120,51 @@ export const LandingPage = () => {
       {showSignInModal && <SignInForBookmarkModal onClose={closeSignInModal} />}
     </>
   );
+};
+
+export const useGetPosts = ({
+  page = 0,
+  roles,
+  investmentMax,
+  investmentMin,
+  series,
+  pathStatus,
+}: {
+  page?: number;
+  roles?: string[];
+  investmentMax?: number;
+  investmentMin?: number;
+  series?: Series[];
+  pathStatus?: number;
+}) => {
+  const { postService } = useGuardContext(ServiceContext);
+
+  const { data: postsData } = useQuery({
+    queryKey: [
+      'postService',
+      'getPosts',
+      page,
+      roles,
+      investmentMax,
+      investmentMin,
+      series,
+      pathStatus,
+    ],
+    queryFn: async () => {
+      const response = await postService.getPosts({
+        page,
+        roles,
+        investmentMax,
+        investmentMin,
+        series,
+        pathStatus,
+      });
+      if (response.type === 'success') {
+        return response.data;
+      }
+      throw new Error('회사 정보를 가져오는데 실패했습니다.');
+    },
+  });
+
+  return { postsData };
 };
